# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KWG Directory API
  version: 0.1.0
  description: OpenChain Korea Work Group 멤버 주소록 REST API. 인증은 next-auth 세션 쿠키 기반. /api/auth/* 는 next-auth 내부 핸들러로 본 문서 범위 밖.
  license:
    name: Apache-2.0
servers:
  - url: /
    description: 동일 오리진
...
paths:
...
  /api/me/export:
    get:
      summary: GDPR — 내 데이터 내보내기(JSON 다운로드). rate-limited.
      tags:
        - me
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 데이터 export(application/json 첨부)
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/me/delete:
    delete:
      summary: GDPR — 계정·데이터 영구 삭제. rate-limited(일 1회).
      tags:
        - me
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeleteRequest'
      responses:
        '204':
          description: 삭제됨(본문 없음)
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/reject:
    post:
      summary: '[admin] 가입 거절(사유 필수).'
      tags:
        - admin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminReject'
      responses:
        '200':
          description: 거절됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/reinvite:
    post:
      summary: '[admin] 메일링 초대 재발송.'
      tags:
        - admin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminId'
      responses:
        '200':
          description: 재발송됨
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/notifications:
    get:
      summary: '[admin] 알림 목록.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 알림 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationList'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: '[admin] 알림 읽음 처리.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 처리됨
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Success:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    AccountDeleteRequest:
      type: object
      properties:
        confirmation:
          type: string
          enum:
            - DELETE-MY-ACCOUNT
          description: 삭제 확인 토큰. 정확히 "DELETE-MY-ACCOUNT" 이어야 함.
      required:
        - confirmation
...
    AdminId:
      type: object
      description: zod AdminIdSchema 와 동기화.
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    AdminReject:
      type: object
      description: zod AdminRejectSchema 와 동기화.
      properties:
        id:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - id
        - reason
...
    NotificationList:
      type: array
      items:
        type: object